+++ src/lib_java/java_classes/Label.java
@@ -7,9 +7,9 @@ import vizdoom.*;
 public class Label{
     public int objectId;
     public String objectName;
-    public int value;
+    public byte value;
 
-    Label(int objectId, String objectName, int value){
+    Label(int objectId, String objectName, byte value){
         this.objectId = objectId;
         this.objectName = objectName;
         this.value = value;
